energy ovs tests significantly simplified and pass

+++ src/collisions/reactions.jl
@@ -68,7 +68,7 @@ abstract type ReactionModel end
 """
 By default, rate_coeff looks for a lookup table stored in the reaction struct
 """
-function rate_coeff(::ReactionModel, rxn::Reaction, energy::Float64)
+function rate_coeff(::ReactionModel, rxn::Reaction, energy)
     ind = Base.trunc(Int64, energy)
     N = length(rxn.rate_coeffs) - 2
     ind = ind > N ? N : ind
